Lead create e index

+++ Models/Site/Lead_atendentes.cs
@@ -299,11 +299,40 @@ namespace gestaoContadorcomvc.Models.Site
 
             try
             {
-                comando.CommandText = "DELETE from lead_atendentes WHERE lead_atendentes.lead_atendentes_id = @lead_atendentes_id and lead_atendentes.lead_atendentes_conta_id = @conta_id;";
-                comando.Parameters.AddWithValue("@conta_id", conta_id);
-                comando.Parameters.AddWithValue("@lead_atendentes_id", lead_atendentes_id);
-                comando.ExecuteNonQuery();
-                Transacao.Commit();
+                MySqlDataReader dr_1;
+                MySqlCommand dr_1_c = conn.CreateCommand();
+                dr_1_c.Connection = conn;
+                dr_1_c.Transaction = Transacao;
+                dr_1_c.CommandText = "SELECT COUNT(l.lead_id) as 'leads' from lead as l WHERE l.lead_lead_atendentes_id = @lead_lead_atendentes_id;";
+                dr_1_c.Parameters.AddWithValue("conta_id", conta_id);
+                dr_1_c.Parameters.AddWithValue("lead_lead_atendentes_id", lead_atendentes_id);                
+                dr_1 = dr_1_c.ExecuteReader();
+
+                int leads = 0;
+
+                if (dr_1.HasRows)
+                {
+                    while (dr_1.Read())
+                    {
+                        leads = Convert.ToInt32(dr_1["leads"]);
+                    }
+                }
+                dr_1.Close();
+
+                if (leads > 0)
+                {
+                    retorno = "Atendente não pode ser excluído, pois existem " + leads + " LEAD´s atribuídos a ele!";                    
+
+                    return retorno;
+                }
+                else
+                {
+                    comando.CommandText = "DELETE from lead_atendentes WHERE lead_atendentes.lead_atendentes_id = @lead_atendentes_id and lead_atendentes.lead_atendentes_conta_id = @conta_id;";
+                    comando.Parameters.AddWithValue("@conta_id", conta_id);
+                    comando.Parameters.AddWithValue("@lead_atendentes_id", lead_atendentes_id);
+                    comando.ExecuteNonQuery();
+                    Transacao.Commit();
+                }
             }
             catch (Exception e)
             {
@@ -318,6 +347,65 @@ namespace gestaoContadorcomvc.Models.Site
             }
 
             return retorno;
+        }
+
+        public Lead_atendentes busca_proximo_atendente(int conta_id, int fila)
+        {
+            Lead_atendentes atendente = new Lead_atendentes();
+
+            conn.Open();
+            MySqlCommand comando = conn.CreateCommand();
+            MySqlTransaction Transacao;
+            Transacao = conn.BeginTransaction();
+            comando.Connection = conn;
+            comando.Transaction = Transacao;
+
+            try
+            {
+                if(fila == 1)
+                {
+                    comando.CommandText = "SELECT MIN(l.lead_atendentes_filaUm), l.lead_atendentes_id, l.lead_atendentes_nome, l.lead_atendentes_celular, l.lead_atendentes_email from lead_atendentes as l WHERE l.lead_atendentes_conta_id = @conta_id and l.lead_atendentes_atende_fila_um = true and l.lead_atendentes_filaUm = (SELECT MIN(l.lead_atendentes_filaUm) from lead_atendentes as l WHERE l.lead_atendentes_conta_id = @conta_id and l.lead_atendentes_atende_fila_um = true);";
+                }
+                if (fila == 2)
+                {
+                    comando.CommandText = "SELECT MIN(l.lead_atendentes_filaDois), l.lead_atendentes_id, l.lead_atendentes_nome, l.lead_atendentes_celular, l.lead_atendentes_email from lead_atendentes as l WHERE l.lead_atendentes_conta_id = @conta_id and l.lead_atendentes_atende_fila_dois = true and l.lead_atendentes_filaDois = (SELECT MIN(l.lead_atendentes_filaDois) from lead_atendentes as l WHERE l.lead_atendentes_conta_id = @conta_id and l.lead_atendentes_atende_fila_dois = true);";
+                }                
+                comando.Parameters.AddWithValue("@conta_id", conta_id);                
+                comando.ExecuteNonQuery();
+
+                var leitor = comando.ExecuteReader();
+
+                if (leitor.HasRows)
+                {
+                    while (leitor.Read())
+                    {
+                        if (DBNull.Value != leitor["lead_atendentes_id"])
+                        {
+                            atendente.lead_atendentes_id = Convert.ToInt32(leitor["lead_atendentes_id"]);
+                        }
+                        else
+                        {
+                            atendente.lead_atendentes_id = 0;
+                        }                        
+                        atendente.lead_atendentes_nome = leitor["lead_atendentes_nome"].ToString();
+                        atendente.lead_atendentes_celular = leitor["lead_atendentes_celular"].ToString();
+                        atendente.lead_atendentes_email = leitor["lead_atendentes_email"].ToString();
+                    }
+                }
+            }
+            catch (Exception)
+            {
+
+            }
+            finally
+            {
+                if (conn.State == System.Data.ConnectionState.Open)
+                {
+                    conn.Close();
+                }
+            }
+
+            return atendente;
         }
